+++ sdks/outpost-typescript/src/models/operations/gettenantdestinationtypeschema.ts
@@ -22,6 +22,7 @@ export const GetTenantDestinationTypeSchemaType = {
   Rabbitmq: "rabbitmq",
   Hookdeck: "hookdeck",
   AwsKinesis: "aws_kinesis",
+  AwsS3: "aws_s3",
 } as const;
 /**
  * The type of the destination.